requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
pydantic==2.4.2
pydantic-settings==2.0.3

# Google Cloud Services
google-generativeai==0.3.1
google-cloud-firestore==2.13.1
google-auth==2.23.4
firebase-admin==6.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Processing & ML
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
sentence-transformers==2.2.2

# Visualization
plotly==5.17.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.0
asyncio==3.4.3
aiofiles==23.2.1
requests==2.31.0

# Development & Testing
pytest==7.4.2
pytest-asyncio==0.21.1
